behave: better sub-project structure
