+++ behave/testlib/__init__.py
@@ -1,0 +1,1 @@
+""" Helper library for Mock's BDD """
